Improve UI

+++ Topics/Generics and Object/Correct the usage of non-generic class/src/Main.java
@@ -1,3 +1,4 @@
+
 class Holder {
     private Object value;
 
@@ -16,7 +17,7 @@ class Main {
         holder.set(256);
 
         // correct the line to make the code compile
-        Integer value = holder.get();
+        Integer value = (Integer) holder.get();
 
         // do not change
         System.out.println(value);
